int -> size_t fixing a warning

+++ test/globalbenchmarks/NavigationBenchmark.cpp
@@ -121,7 +121,7 @@ int main(int argc, char* argv[])
     visualizer.AddVolume( *world );
 
     Vector<Daughter> const* daughters = world->GetLogicalVolume()->GetDaughtersp();
-    for(int i=0; i<daughters->size(); ++i) {
+    for(size_t i=0; i<daughters->size(); ++i) {
       VPlacedVolume const* daughter = (*daughters)[i];
       Transformation3D const& trf1 = *(daughter->GetTransformation());
       visualizer.AddVolume(*daughter, trf1);
